Impl Polygon convexity function on the type
The convexity function was previously only available within the
Extremes module, but it's generally useful and will be required by
other algorithms. This moves the function and implements it directly
on the type.
